Add signature support for Django 2.2+ constraints.
Constraints are a feature available in Django 2.2+. They're a way of
listing explicit constraints on a table, for ensuring that a set of
fields are unique (potentially with a conditional query that must also
be met), or for ensuring that a condition passes when inserting data.
These are dependent on the support in the database backend.

This initial change introduces signature support for
serializing/deserializing information on conditions specified on a
model. They're not currently used as part of the evolution process, but
that will be arriving in an upcoming change.

Testing Done:
Unit tests pass for all versions of Python and Django.

Reviewed at https://reviews.reviewboard.org/r/11119/