Rewrite the README and part of the introduction to the docs.
The Django Evolution README was horribly out-of-date. It didn't make a
good case for itself in the modern world, and gave the impression the
project hasn't been touched in many years.

This change completely rewrites the README. It's now in
ReStructuredText, instead of Markdown, and included as package metadata.
The new content talks about where Django Evolution is useful and touches
upon the compatibility with Django's migrations. It includes updated
compatibility information and a link to the (upcoming) public docs, plus
a blurb about its usage in Review Board.

Testing Done:
Checked the rendered README and the new docs.

Built a package and verified the presence of the README in the metadata.

Reviewed at https://reviews.reviewboard.org/r/11123/