Allow creation of a MigrationList from a list of migration names.
The new `MigrationList.from_names()` allows for easy creation of a
`MigrationList` from just an app label and a set of migration names,
without any need to inspect the database or a signature. This will
be used in upcoming changes that need to more carefully build and
operate on lists of migrations.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11222/