Add a new evolutions argument for some signal emissions.
The `apply_evolutions` and `applied_evolutions` signals now provide an
`evolutions` argument, which is a list of evolutions that will be/were
applied. This is currently equivalent to accessing
`task.new_evolutions`, but will be important for the upcoming work to
add dependency relations between migrationss and evolutions.

The `evolve` management command takes advantage of this new argument
when the verbosity is 2 or higher, listing the specific evolutions that
are part of the signal emission.

Testing Done:
Unit tests passed.

Ran an evolution at each verbosity level (0, 1, 2, 3) and saw the
expected output for that level

Reviewed at https://reviews.reviewboard.org/r/11223/