Allow providing explicit SQL and evolutions to EvolveAppTask.execute().
`EvolveAppTask.execute()` now allows the caller to provide specific SQL
to execute (instead of the task-prepared SQL) and a list of evolutions
for signal emissions (instead of the task-prepared `new_evolutions`).

This will be used by some upcoming work to enable finer-grained batches
of evolutions and SQL statements that are calculated independently of
the task's full lists of evolutions and SQL.

Testing Done:
Unit tests passed.

Tested this with in-progress changes that provide explicit SQL and
evolutions.

Reviewed at https://reviews.reviewboard.org/r/11224/