Allow get_app_pending_mutations to take precomputed values.
`get_app_pending_mutations()` is responsible for looking up the
evolutions, stored signature, and current signature for an app, and then
filtering a list of mutations based on the differences between those
signatures. This requires lookups to be performed, which can be slow and
may not reflect the state we want used for the computations.

This change allows the project signatures and the list of mutations to
be provided by the caller instead of automatically looking them up.

It also adds unit tests for this function, since those didn't exist
before.

Testing Done:
Unit tests pass for all versions of Python and Django.

Reviewed at https://reviews.reviewboard.org/r/11249/