Fix regressions from Django 3.2.
Django 3.2 was just released, and triggered some regressions with
Postgres and for index generation.

The index generation issue was just a matter of converting a positional
argument to a keyword argument. This change was backwards-compatible to
all supported versions of Django.

The Postgres issue was caused by the Postgres backend requiring access
to a field's model when renaming a field. We weren't setting the model
on these fields before, and there was no good reason for this. We now
set the model, fixing this issue.

Testing Done:
Unit tests pass on all supported versions of Django, and with all
database types.

Reviewed at https://reviews.reviewboard.org/r/11580/