Lazily fetch the django_evolution app.
After a `syncdb` or `migrate`, `django_evolution.management` attempts to
record new state in the database, but it only does so once the
`django_evolution` app and its tables have been made available in the
database. For this, it needs a reference to the app, which was being
done at module initialization.

It's possible for this code to run too early, before other apps are
properly set up, which could cause failures either fetching the app or
from some logic within the app.

To avoid this, we now lazily fetch the app the first time we need it,
caching it for future reference.

Testing Done:
Unit tests pass.

Successfully created new databases and upgraded existing ones.

Reviewed at https://reviews.reviewboard.org/r/11875/