Update package metadata and tox for supported Python versions.
This updates the classifiers for the package to list Python 3.9 and 3.10
support, and to include metadata that will help ensure `pip` installs
the right package based on the current version of Python in the future,
which will matter as we end up dropping older versions down the road.

The tox configuration has also been updated to test Python 3.10, and to
use compatible release clauses for the right versions of Django.

Testing Done:
Successfully built the package.

Tox successfully ran all unit tests.

Reviewed at https://reviews.reviewboard.org/r/11994/